// src/payload-types.ts
export interface Media {
  id: number;
  alt?: string | null;
  caption?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  prefix?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    square?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    small?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    medium?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    large?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    xlarge?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    og?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "pages".
 */
export interface Page {
  id: number;
  title: string;
  hero: {
    type: 'none' | 'highImpact' | 'mediumImpact' | 'lowImpact' | 'homeHero' | 'splitHeroWithCTA';
    title?: string | null;
    subtitle?: string | null;
    description?: string | null;
    richText?: {
      root: {
        type: string;
        children: {
          type: string;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    } | null;
    links?:
      | {
          link: {
            type?: ('reference' | 'custom') | null;
            newTab?: boolean | null;
            reference?: {
              relationTo: 'pages';
              value: number | Page;
            } | null;
            url?: string | null;
            label: string;
            /**
             * Choose how the link should be rendered.
             */
            appearance?: ('default' | 'outline') | null;
          };
          id?: string | null;
        }[]
      | null;
    missionTitle?: string | null;
    missionText?: string | null;
    missionImage?: (number | null) | Media;
    missionBackgroundColor?:
      | ('bg-beige' | 'bg-black' | 'bg-cream' | 'bg-gray' | 'bg-mustard' | 'bg-purple' | 'bg-rust' | 'bg-white')
      | null;
    button?: {
      text?: string | null;
      link?: string | null;
      variant?: ('default' | 'mustard' | 'link') | null;
      size?: ('default' | 'small' | 'large') | null;
      fullWidth?: boolean | null;
    };
    media?: (number | null) | Media;
    splitHero?: {
      heading?: string | null;
      mainTitle?: string | null;
      mainSubtitle?: string | null;
      mainDescription?: {
        root: {
          type: string;
          children: {
            type: string;
            version: number;
            [k: string]: unknown;
          }[];
          direction: ('ltr' | 'rtl') | null;
          format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
          indent: number;
          version: number;
        };
        [k: string]: unknown;
      } | null;
      contentPosition?: ('left' | 'right') | null;
      contentType?: ('card' | 'image' | 'none') | null;
      contentImage?: (number | null) | Media;
      useCardBlock?: boolean | null;
      cardBlock?: CardBlock[] | null;
    };
  };
  layout: (
    | BannerBlockConfig
    | CardBlock
    | ContentBlock
    | CallToActionBlock
    | CustomMediaBlockConfig
    | DescriptionBlockConfig
    | EventListConfig
    | FAQBlock
    | FeatureBlockConfig
    | FleetListConfig
    | FormBlock
    | HotelsBlockConfig
    | HotelServicingConfig
    | HTMLBlockConfig
    | InfoBlockConfig
    | LocalReviewsConfig
    | MediaBlock
    | NewsletterBlock
    | RatingBlock
    | RequestQuoteConfig
    | ShopperApprovedBlockConfig
    | TestimonialBlockConfig
    | ToursListConfig
    | TrustedPartnerConfig
    | VisualChecklistBlock
  )[];
  meta?: {
    title?: string | null;
    /**
     * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
     */
    image?: (number | null) | Media;
    description?: string | null;
  };
  /**
   * Toggle this to enable or disable SEO metadata for this page. When disabled, the page will not be indexed by search engines.
   */
  enableMetadata?: boolean | null;
  publishedAt?: string | null;
  slug?: string | null;
  slugLock?: boolean | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "CardBlock".
 */
export interface CardBlock {
  title: string;
  subtitle?: string | null;
  primaryText?: string | null;
  primaryTextColor?: ('text-white' | 'text-beige' | 'text-black' | 'text-mustard' | 'text-gray' | 'text-rust') | null;
  /**
   * Use [highlight] text [/highlight] syntax to highlight specific words or phrases
   */
  secondaryText?: string | null;
  secondaryTextColor?: ('text-white' | 'text-beige' | 'text-black' | 'text-mustard' | 'text-gray' | 'text-rust') | null;
  media?: (number | null) | Media;
  imagePosition?: ('left' | 'right') | null;
  button?: {
    text?: string | null;
    link?: string | null;
    variant?: ('default' | 'mustard' | 'link') | null;
    size?: ('default' | 'small' | 'large') | null;
    fullWidth?: boolean | null;
  };
  backgroundColor?: ('bg-white' | 'bg-mustard') | null;
  titleColor?: ('text-mustard' | 'text-gray') | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'cardBlock';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "BannerBlockConfig".
 */
export interface BannerBlockConfig {
  style: 'info' | 'warning' | 'error' | 'success';
  content: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: 'bannerBlock';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ContentBlock".
 */
export interface ContentBlock {
  columns?:
    | {
        size?: ('oneThird' | 'half' | 'twoThirds' | 'full') | null;
        richText?: {
          root: {
            type: string;
            children: {
              type: string;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        } | null;
        enableLink?: boolean | null;
        link?: {
          type?: ('reference' | 'custom') | null;
          newTab?: boolean | null;
          reference?: {
            relationTo: 'pages';
            value: number | Page;
          } | null;
          url?: string | null;
          label: string;
          /**
           * Choose how the link should be rendered.
           */
          appearance?: ('default' | 'outline') | null;
        };
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'content';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "CallToActionBlock".
 */
export interface CallToActionBlock {
  richText?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  links?:
    | {
        link: {
          type?: ('reference' | 'custom') | null;
          newTab?: boolean | null;
          reference?: {
            relationTo: 'pages';
            value: number | Page;
          } | null;
          url?: string | null;
          label: string;
          /**
           * Choose how the link should be rendered.
           */
          appearance?: ('default' | 'outline') | null;
        };
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'cta';
}
// ...
export interface HotelsBlockConfig {
  relationTo?: 'hotels' | null;
  limit?: number | null;
  selectedDocs?:
    | {
        relationTo: 'hotels';
        value: number | Hotel;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'hotelsBlock';
}
// ...
export interface Header {
  id: number;
  navItems?:
    | {
        link: {
          type?: ('reference' | 'custom') | null;
          newTab?: boolean | null;
          reference?: {
            relationTo: 'pages';
            value: number | Page;
          } | null;
          url?: string | null;
          label: string;
        };
        subItems?:
          | {
              link: {
                type?: ('reference' | 'custom') | null;
                newTab?: boolean | null;
                reference?: {
                  relationTo: 'pages';
                  value: number | Page;
                } | null;
                url?: string | null;
                label: string;
              };
              id?: string | null;
            }[]
          | null;
        id?: string | null;
      }[]
    | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "footer".
 */
export interface Footer {
  id: number;
  logo: number | Media;
  address?: string | null;
  phone?: string | null;
  email?: string | null;
  faqLink: {
    link: {
      type?: ('reference' | 'custom') | null;
      newTab?: boolean | null;
      reference?: {
        relationTo: 'pages';
        value: number | Page;
      } | null;
      url?: string | null;
      label: string;
    };
  };
  socialLinks?:
    | {
        label?: string | null;
        newTab?: boolean | null;
        url?: string | null;
        icon?: (number | null) | Media;
        id?: string | null;
      }[]
    | null;
  legalLinks?:
    | {
        link: {
          type?: ('reference' | 'custom') | null;
          newTab?: boolean | null;
          reference?: {
            relationTo: 'pages';
            value: number | Page;
          } | null;
          url?: string | null;
          label: string;
        };
        id?: string | null;
      }[]
    | null;
  websiteCredit?: string | null;
  websiteCreditUrl?: string | null;
  navItems?:
    | {
        link: {
          type?: ('reference' | 'custom') | null;
          newTab?: boolean | null;
          reference?: {
            relationTo: 'pages';
            value: number | Page;
          } | null;
          url?: string | null;
          label: string;
        };
        id?: string | null;
      }[]
    | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}